pyproject.toml
[tool.black]
line-length = 79
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.poetry]
name = "post-from-telegram-to-wordpress"
version = "0.1.0"
description = "create a post on wordpress by sending message to the telegram-bot or chat"
authors = ["Serhii Hrekov <themebeat@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
python-telegram-bot = "13.7.0"
python-dotenv = "^1.0.1"
python-wordpress = "^0.1.3"
requests = "2.26.0"
pydantic = "^2.6.4"
aiohttp = "^3.9.3"
uvicorn = "^0.29.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
